fixes #3585
git-svn-id: https://devel.scolab.eu/gvsig-online/gvsigonline_v2/tags/gvsigol_v2.3.5@5119 216424d3-d1a8-4ca4-951f-a609719bba38

+++ gvsigol/gvsigol_services/views.py
@@ -423,7 +423,8 @@ def layer_delete(request, layer_id):
 @staff_required
 def layer_delete_operation(request, layer_id):
     layer = Layer.objects.get(pk=layer_id)
-    mapservice_backend.deleteGeoserverLayerGroup(layer.layer_group)
+    if layer.layer_group.name != '__default__':
+        mapservice_backend.deleteGeoserverLayerGroup(layer.layer_group)
     mapservice_backend.deleteResource(layer.datastore.workspace, layer.datastore, layer)
     mapservice_backend.deleteLayerStyles(layer)
     signals.layer_deleted.send(sender=None, layer=layer)
